first version

+++ src/main/java/com/sci/ro/finalproject/finalproject/repository/TripRepository.java
@@ -3,5 +3,5 @@ package com.sci.ro.finalproject.finalproject.repository;
 import com.sci.ro.finalproject.finalproject.model.Trip;
 import org.springframework.data.repository.CrudRepository;
 
-public interface TripRepository extends CrudRepository<Trip, Integer> {
+public interface TripRepository extends CrudRepository<Trip, Long> {
 }
